split deployment in two components + layout fixes
