+++ frontend/src/App.js
@@ -7,7 +7,8 @@ import { LoginPanel } from './components/loginPanel';
 import { FormList } from './components/forms/formList';
 import { UploadForm } from './components/forms/uploadForm';
 import { ArchivedForms } from './components/forms/archivedList';
-import { UploadDeployment } from './components/uploadDeployment';
+import { UploadDeployment } from './components/deployments/upload';
+import { SelectCreateDeployment } from './components/deployments/selectCreate';
 import './App.css';
 
 
@@ -20,7 +21,8 @@ class App extends Component {
         <Route exact path="/login" component={LoginPanel} />
         <Route exact path="/upload-form" component={UploadForm} />
         <Route exact path="/archived-forms" component={ArchivedForms} />
-        <Route exact path="/deployments" component={UploadDeployment} />
+        <Route exact path="/deployments" component={SelectCreateDeployment} />
+        <Route exact path="/deployments/:deployment" component={UploadDeployment} />
         <Route exact path="/submissions/:formId" component={SubmissionList} />
       </div>
     );
